server/pyproject.toml
[project]
name = "couch-server"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "websockets>=14.0",
    "pyyaml>=6.0",
    "faster-whisper>=1.0.0",
    "silero-vad>=5.1",
    "torch>=2.0.0",
    "numpy>=1.26.0",
    "ollama>=0.4.0",
    "playwright>=1.44.0",
    "anthropic>=0.40.0",
    "pydantic-settings>=2.0.0",
    "pillow>=10.0.0",
]

[tool.uv]
package = false
